final_project/project/requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
cryptography==38.0.4
Faker==15.3.4
pydantic==1.10.2
PyMySQL==1.0.2
pytest==7.2.0
pytest-xdist==3.1.0
requests==2.28.1
selenium==4.6.1
SQLAlchemy==1.4.44
webdriver-manager==3.8.5

